cmStandardLevelResolver: Added to handle standard level queries
Refactored out of cmMakefile
